+++ src/dipdup/datasources/datasource.py
@@ -6,6 +6,8 @@ from typing import Optional
 from typing import Set
 from typing import Tuple
 
+from aiohttp.hdrs import METH_GET
+
 from dipdup.config import HTTPConfig
 from dipdup.datasources.subscription import HeadSubscription
 from dipdup.datasources.subscription import Subscription
@@ -36,6 +38,26 @@ class Datasource(HTTPGateway):
 
     def set_logger(self, name: str) -> None:
         self._logger = FormattedLogger(self._logger.name, name + ': {}')
+
+
+class HttpDatasource(Datasource):
+    _default_http_config = HTTPConfig(
+        cache=True,
+        retry_count=5,
+        retry_sleep=1,
+        ratelimit_rate=0,
+        ratelimit_period=0,
+    )
+
+    def __init__(self, url: str, http_config: Optional[HTTPConfig] = None) -> None:
+        super().__init__(url, self._default_http_config.merge(http_config))
+        self._logger = _logger
+
+    async def get(self, url: str, cache: bool = False, weight: int = 1, **kwargs):
+        return await self.request(METH_GET, url, cache, weight, **kwargs)
+
+    async def run(self) -> None:
+        pass
 
 
 # TODO: Generic interface